+++ Changelog.md
@@ -6,7 +6,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-* Added support for 15- and 24-bit colour depth and 8-bit colour maps on VNC
 
 ### Changed
 
@@ -17,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## [v0.6.0] - 2020-06-29
+### Added
+* Added support for 15- and 24-bit colour depth and 8-bit colour maps on VNC
 
 ## [v0.5.0] - 2020-06-22
 ### Added